requirements.txt
Flask>=3.0.1,<3.1.0          # Minimum version 3.0.1 and less than 3.1.0
bson                          # Use the bson included with pymongo
python-dotenv>=1.0.0,<1.1.0  # Minimum version 1.0.0 and less than 1.1.0
pymongo>=4.6.2,<4.7.0        # Minimum version 4.6.2 and less than 4.7.0
gunicorn>=21.1.0,<21.2.0     # Minimum version 21.1.0 and less than 21.2.0
requests>=2.30.0,<2.31.0     # Minimum version 2.30.0 and less than 2.31.0
vault-cli>=2.2.0,<2.3.0      # Minimum version 2.2.0 and less than 2.3.0
flask-cors>=4.0.0,<5.0.0

# Flask → Similar to express

# python-dotenv → Equivalent to dotenv

# pymongo → Equivalent to mongoose

# gunicorn → Equivalent to nodemon for running servers

# requests → For making API requests (like using Axios in Node.js)

# vault-cli → Equivalent to node-vault for interacting with HashiCorp Vault